Fix deploy scripts, add travis ssh keygen and add heroku conf

+++ app/templates/scripts/gh-pages.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+if [ ! -z "$BRANCH" ] && [ "$BRANCH" != "$TRAVIS_BRANCH" ]; then
+	echo -e "\033[33m Not for this $TRAVIS_BRANCH branch."
+	exit 1
+fi
 
 if [ -z "$REPO" ]
 then
@@ -11,6 +15,11 @@ else
 fi
 message=$(git log -1 --pretty=%B)
 
+if [ ! -d "dist" ]; then
+	echo -e "\033[31m You must run \"grunt build\" before commit to gh-pages."
+	exit 1
+fi
+
 cd dist
 rm -rf .git
 git init
@@ -19,5 +28,5 @@ git push origin --delete gh-pages > /dev/null 2>&1
 git checkout -b gh-pages
 git add --all
 git commit -m "$message"
-git push origin gh-pages > /dev/null 2>&1
+git push origin gh-pages -f > /dev/null 2>&1
 cd ..
